refactor duplicated code and document changes
