+++ packages/defiscan-frontend/src/pages/review/views/explorer/AdminsTab.tsx
@@ -1,15 +1,15 @@
-import { useState, useMemo, useRef, useLayoutEffect, useEffect } from 'react'
+import { useState, useMemo } from 'react'
 import { Badge } from '../../../../components/Badge'
 import { AddressDisplay } from '../../../../components/AddressDisplay'
 import { UsdValue } from '../../../../components/UsdValue'
 import { formatUsdValue } from '../../../../utils/format'
 import { getHumanAdmins } from '../../../../utils/admins'
+import type { CompiledReview, CompiledAdmin } from '../../../../types'
 import {
-  type CompiledReview,
-  type CompiledAdmin,
-  type Mitigation,
-} from '../../../../types'
-import { MitigationBadge } from '../../../../components/MitigationBadge'
+  SortHeader,
+  MitigationsSummary,
+  ExpandedAdminFunctions,
+} from './shared'
 
 interface AdminsTabProps {
   review: CompiledReview
@@ -251,7 +251,7 @@ function AdminRow({
           )}
         </td>
         <td className="px-4 py-2.5">
-          <AdminMitigationsSummary admin={admin} />
+          <MitigationsSummary functions={admin.functions} />
         </td>
         <td className="px-4 py-2.5 text-right font-medium text-text-primary">
           {admin.functions.length}
@@ -259,96 +259,12 @@ function AdminRow({
       </tr>
       {isExpanded && (
         <tr>
-          <td colSpan={7} className="px-0 py-0">
-            <ExpandedFunctions admin={admin} />
+          <td colSpan={6} className="px-0 py-0">
+            <ExpandedAdminFunctions admin={admin} />
           </td>
         </tr>
       )}
     </>
-  )
-}
-
-function ExpandedFunctions({ admin }: { admin: CompiledAdmin }) {
-  return (
-    <div className="bg-bg-muted/50 border-t border-border">
-      {admin.description && (
-        <p className="px-6 py-3 text-sm text-text-secondary border-b border-border/50 leading-relaxed">
-          {admin.description}
-        </p>
-      )}
-      <div className="px-6 py-3">
-        <table className="w-full text-xs">
-          <thead>
-            <tr className="text-text-muted">
-              <th className="text-left pb-1 font-medium">Contract</th>
-              <th className="text-left pb-1 font-medium">Function</th>
-              <th className="text-left pb-1 font-medium">Mitigations</th>
-              <th className="text-right pb-1 font-medium">TVL</th>
-              <th className="text-right pb-1 font-medium">
-                Reachable Contracts
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {admin.functions.map((fn) => (
-              <tr
-                key={`${fn.contractAddress}-${fn.functionName}`}
-                className="border-t border-border/30"
-              >
-                <td className="py-1.5 text-text-secondary">
-                  {fn.contractName}
-                </td>
-                <td className="py-1.5">
-                  <span className="font-mono text-text-primary">
-                    {fn.functionName}()
-                  </span>
-                </td>
-                <td className="py-1.5">
-                  {fn.mitigations && fn.mitigations.length > 0 ? (
-                    <div className="flex flex-wrap gap-0.5">
-                      {fn.mitigations.map((m, i) => (
-                        <MitigationBadge key={i} mitigation={m} />
-                      ))}
-                    </div>
-                  ) : (
-                    <span className="text-text-muted">-</span>
-                  )}
-                </td>
-                <td className="py-1.5 text-right tabular-nums">
-                  {fn.directFundsUsd > 0 ? (
-                    <span className="text-capital font-medium">
-                      {formatUsdValue(fn.directFundsUsd)}
-                    </span>
-                  ) : (
-                    <span className="text-text-muted">-</span>
-                  )}
-                </td>
-                <td className="py-1.5 text-right">
-                  {fn.reachableContracts.length > 0 ? (
-                    <span className="text-text-primary">
-                      {fn.reachableContracts.length}
-                      {fn.reachableContracts.some((rc) => rc.fundsAtRisk) && (
-                        <span className="ml-1 text-capital">
-                          (
-                          {formatUsdValue(
-                            fn.reachableContracts
-                              .filter((rc) => rc.fundsAtRisk)
-                              .reduce((s, rc) => s + rc.fundsUsd, 0),
-                          )}
-                          )
-                        </span>
-                      )}
-                    </span>
-                  ) : (
-                    <span className="text-text-muted">-</span>
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
   )
 }
 
@@ -364,150 +280,3 @@ function AdminsSummaryLabel({ admins }: { admins: CompiledAdmin[] }) {
     </span>
   )
 }
-
-function deduplicateMitigations(mitigations: Mitigation[]): Mitigation[] {
-  const seen = new Set<string>()
-  const result: Mitigation[] = []
-  for (const m of mitigations) {
-    const key = `${m.type}:${m.delaySeconds ?? ''}:${m.valueRange?.min ?? ''}:${m.valueRange?.max ?? ''}:${m.relativeValue?.maxChangePercent ?? ''}:${m.description}`
-    if (!seen.has(key)) {
-      seen.add(key)
-      result.push(m)
-    }
-  }
-  return result
-}
-
-function AdminMitigationsSummary({ admin }: { admin: CompiledAdmin }) {
-  const allMitigations: Mitigation[] = []
-  for (const fn of admin.functions) {
-    if (fn.mitigations) {
-      allMitigations.push(...fn.mitigations)
-    }
-  }
-
-  const unique = deduplicateMitigations(allMitigations)
-  const measureRef = useRef<HTMLDivElement>(null)
-  const [visibleCount, setVisibleCount] = useState(unique.length)
-
-  const measure = () => {
-    const measureDiv = measureRef.current
-    if (!measureDiv || unique.length === 0) return
-    const td = measureDiv.closest('td')
-    if (!td) return
-    // Available width = td content width minus padding (px-4 = 16px * 2)
-    const available = td.clientWidth - 32
-    const reservedForLabel = 28
-    const children = Array.from(
-      measureDiv.querySelectorAll<HTMLElement>('[data-measure]'),
-    )
-    let used = 0
-    let count = 0
-    for (const child of children) {
-      used += child.offsetWidth + (count > 0 ? 2 : 0) // 2px gap
-      if (used <= available - reservedForLabel) {
-        count++
-      } else {
-        break
-      }
-    }
-    if (count === unique.length) {
-      setVisibleCount(unique.length)
-    } else {
-      setVisibleCount(Math.max(count, 1))
-    }
-  }
-
-  // Initial measurement before paint
-  useLayoutEffect(measure, [unique.length])
-
-  // Re-measure when the parent td resizes (zoom, column changes).
-  // Observing the td (not our container) avoids feedback loops since
-  // hiding badges doesn't change the td's width.
-  useEffect(() => {
-    const measureDiv = measureRef.current
-    if (!measureDiv) return
-    const td = measureDiv.closest('td')
-    if (!td) return
-    const observer = new ResizeObserver(measure)
-    observer.observe(td)
-    return () => observer.disconnect()
-  }, [unique.length])
-
-  if (unique.length === 0) {
-    return <span className="text-text-muted">-</span>
-  }
-
-  const remaining = unique.length - visibleCount
-
-  return (
-    <div className="relative">
-      {/* Hidden measurement layer — always renders ALL badges for stable measurement */}
-      <div
-        ref={measureRef}
-        aria-hidden
-        className="flex flex-nowrap gap-0.5 items-center invisible absolute top-0 left-0 pointer-events-none"
-      >
-        {unique.map((m, i) => (
-          <span key={i} data-measure className="shrink-0">
-            <MitigationBadge mitigation={m} />
-          </span>
-        ))}
-      </div>
-      {/* Visible badges */}
-      <div className="flex flex-nowrap gap-0.5 items-center">
-        {unique.slice(0, visibleCount).map((m, i) => (
-          <span key={i} className="shrink-0">
-            <MitigationBadge mitigation={m} />
-          </span>
-        ))}
-        {remaining > 0 && (
-          <span
-            className="shrink-0 text-text-muted text-[10px] leading-4 ml-0.5"
-            title={`${unique.length} unique mitigations total`}
-          >
-            +{remaining}
-          </span>
-        )}
-      </div>
-    </div>
-  )
-}
-
-function SortHeader({
-  field,
-  label,
-  current,
-  dir,
-  onClick,
-  className,
-}: {
-  field: SortField
-  label: string
-  current: SortField
-  dir: SortDir
-  onClick: (f: SortField) => void
-  className?: string
-}) {
-  const isActive = current === field
-  return (
-    <th
-      className={`px-4 py-2 font-medium text-text-secondary cursor-pointer select-none hover:text-text-primary transition-colors text-left ${className ?? ''}`}
-      onClick={() => onClick(field)}
-    >
-      <span className="inline-flex items-center gap-1">
-        {label}
-        {isActive && (
-          <svg className="w-3 h-3" viewBox="0 0 12 12" fill="currentColor">
-            {dir === 'desc' ? (
-              <path d="M6 8L2 4h8z" />
-            ) : (
-              <path d="M6 4l4 4H2z" />
-            )}
-          </svg>
-        )}
-      </span>
-    </th>
-  )
-}
-
